+++ save_config.sh
@@ -1,1 +1,2 @@
-cp /etc/graphit-tool.conf /etc/graphit-tool.rpmsave
+test -f /etc/graphit-tool.conf && cp /etc/graphit-tool.conf /etc/graphit-tool.rpmsave
+exit 0
